requirements.txt
celery==5.2.6
channels==3.0.4
Django==3.2.6
django_pandas==0.6.6
djangorestframework==3.13.1
matplotlib==3.3.3
missingno==0.4.2
numpy==1.22.3
pandas==1.2.0
redis==3.5.3
scikit_learn==1.0.2
seaborn==0.11.1
xgboost==1.3.3
